# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tokenization Platform API",
        "description": "Documentação da API da plataforma de tokenização",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/investments/purchase": {
            "post": {
                "tags": [
                    "Investments"
                ],
                "summary": "Comprar tokens",
                "description": "Realizar compra de tokens de um imóvel.",
                "operationId": "b1517d264a3b903db97db0e95406d342",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/investments/history": {
            "get": {
                "tags": [
                    "Investments"
                ],
                "summary": "Histórico de investimentos",
                "description": "Histórico de investimentos do usuário autenticado.",
                "operationId": "e4914f0002d40b85e4142c31370ed60f",
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/transacoes-financeiras": {
            "get": {
                "tags": [
                    "Financial Transactions"
                ],
                "summary": "Listar transações",
                "description": "Listar transações financeiras.",
                "operationId": "2376f46ec6613b2db7711d618e2e4738",
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Financial Transactions"
                ],
                "summary": "Registrar transação",
                "description": "Criar nova transação financeira.",
                "operationId": "bed038e4a43be1a19be4c2b7af3d407c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_investidor",
                                    "tipo",
                                    "valor",
                                    "data_transacao"
                                ],
                                "properties": {
                                    "id_investidor": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "deposito"
                                    },
                                    "valor": {
                                        "type": "number",
                                        "example": 1000
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pendente"
                                    },
                                    "referencia": {
                                        "type": "string",
                                        "example": "PIX"
                                    },
                                    "data_transacao": {
                                        "type": "string",
                                        "example": "2024-01-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/transacoes-financeiras/{id}": {
            "get": {
                "tags": [
                    "Financial Transactions"
                ],
                "summary": "Detalhes da transação",
                "description": "Mostrar uma transação específica.",
                "operationId": "3249c2f5ac721198d80a8aa65ac21afc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Financial Transactions"
                ],
                "summary": "Atualizar transação",
                "description": "Atualizar transação financeira.",
                "operationId": "0693fd2a57fa4236bd5b9da44a72f69e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Financial Transactions"
                ],
                "summary": "Excluir transação",
                "description": "Remover transação financeira.",
                "operationId": "0dcc4247a76d7dfb705a828d25abfcdb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/wallet": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Consultar saldo",
                "description": "Exibe saldo da carteira.",
                "operationId": "b0aeab5a5b4b667ad1a4726d02a324cf",
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/wallet/add-funds": {
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Adicionar fundos",
                "description": "Adicionar fundos à carteira.",
                "operationId": "4b51c0e83ec51785cafd358407a852c7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "valor"
                                ],
                                "properties": {
                                    "valor": {
                                        "type": "number",
                                        "example": 500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/wallet/withdraw": {
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Sacar fundos",
                "description": "Solicitar saque de fundos.",
                "operationId": "73405c27b66fe831f83304cefd43c831",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "valor"
                                ],
                                "properties": {
                                    "valor": {
                                        "type": "number",
                                        "example": 200
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}